fix: code redesigning
