+++ src/v0/destinations/iterable/config.js
@@ -20,6 +20,7 @@ const ConfigCategory = {
     name: 'IterableIdentifyConfig',
     action: 'identify',
     endpoint: `users/update`,
+    bulkEndpoint: 'users/bulkUpdate',
   },
   PAGE: {
     name: 'IterablePageConfig',
@@ -35,6 +36,7 @@ const ConfigCategory = {
     name: 'IterableTrackConfig',
     action: 'track',
     endpoint: `events/track`,
+    bulkEndpoint: 'events/trackBulk',
   },
   TRACK_PURCHASE: {
     name: 'IterableTrackPurchaseConfig',
@@ -76,6 +78,10 @@ const constructEndpoint = (dataCenter, category) => {
   return `${baseUrl}${category.endpoint}`;
 };
 
+const BULK_ENDPOINTS = Object.values(ConfigCategory)
+  .filter((config) => config.bulkEndpoint)
+  .map((config) => `/api/${config.bulkEndpoint}`);
+
 const IDENTIFY_MAX_BATCH_SIZE = 1000;
 const IDENTIFY_MAX_BODY_SIZE_IN_BYTES = 4000000;
 
@@ -105,4 +111,5 @@ module.exports = {
   IDENTIFY_MAX_BATCH_SIZE,
   IDENTIFY_MAX_BODY_SIZE_IN_BYTES,
   API_RESPONSE_PATHS,
+  BULK_ENDPOINTS,
 };
