+++ src/v1/destinations/iterable/commonStrategy.js
@@ -1,0 +1,23 @@
+const { ResponseStrategy } = require('./responseStrategy');
+
+class CommonStrategy extends ResponseStrategy {
+  handleSuccess(responseParams) {
+    const { destinationResponse, rudderJobMetadata } = responseParams;
+    const { status } = destinationResponse;
+
+    const responseWithIndividualEvents = rudderJobMetadata.map((metadata) => ({
+      statusCode: status,
+      metadata,
+      error: 'success',
+    }));
+
+    return {
+      status,
+      message: '[ITERABLE Response Handler] - Request Processed Successfully',
+      destinationResponse,
+      response: responseWithIndividualEvents,
+    };
+  }
+}
+
+module.exports = { CommonStrategy };
